requirements.txt
numpy==2.4.2
python-dotenv==1.2.1
python-vlc==3.0.21203
